Agent: Move AgentBinaryHTTPRequestHandler next to the request handlers
