+++ monkey/infection_monkey/exploit/http_agent_binary_request_handler.py
@@ -1,17 +1,23 @@
 import threading
 from http import HTTPStatus
+from http.server import BaseHTTPRequestHandler
 from typing import Callable, Mapping
 
 from common.types import Lock
 from infection_monkey.exploit import IAgentBinaryRepository, RetrievalError
 from infection_monkey.utils.script_dropper import build_bash_dropper
 
-from .agent_binary_request import (
-    AgentBinaryHTTPRequestHandler,
-    AgentBinaryRequest,
-    RequestID,
-    RequestType,
-)
+from .agent_binary_request import AgentBinaryRequest, RequestID, RequestType
+
+
+class AgentBinaryHTTPRequestHandler(BaseHTTPRequestHandler):
+    @classmethod
+    def register_request(cls, request: AgentBinaryRequest):
+        raise NotImplementedError()
+
+    @classmethod
+    def deregister_request(cls, request_id: RequestID):
+        raise NotImplementedError()
 
 
 def get_threading_http_handler(agent_binary_repository: IAgentBinaryRepository):
