+++ monkey/infection_monkey/exploit/http_agent_binary_server.py
@@ -13,12 +13,8 @@ from infection_monkey.network import TCPPortSelector
 from infection_monkey.network.tools import get_interface_to_target
 from infection_monkey.utils.threading import create_daemon_thread
 
-from .agent_binary_request import (
-    AgentBinaryHTTPRequestHandler,
-    AgentBinaryRequest,
-    RequestID,
-    RequestType,
-)
+from .agent_binary_request import AgentBinaryRequest, RequestID, RequestType
+from .http_agent_binary_request_handler import AgentBinaryHTTPRequestHandler
 
 logger = logging.getLogger(__name__)
