+++ monkey/tests/unit_tests/infection_monkey/exploit/test_http_agent_binary_request_handler.py
@@ -10,13 +10,11 @@ import requests
 
 from common import OperatingSystem
 from infection_monkey.exploit import IAgentBinaryRepository, RetrievalError
-from infection_monkey.exploit.agent_binary_request import (
+from infection_monkey.exploit.agent_binary_request import AgentBinaryRequest, RequestID, RequestType
+from infection_monkey.exploit.http_agent_binary_request_handler import (
     AgentBinaryHTTPRequestHandler,
-    AgentBinaryRequest,
-    RequestID,
-    RequestType,
+    get_http_handler,
 )
-from infection_monkey.exploit.http_agent_binary_request_handler import get_http_handler
 
 AGENT_BINARY = b"agent_binary"
 IP = "127.0.0.1"
