+++ monkey/tests/unit_tests/infection_monkey/exploit/test_http_agent_binary_server.py
@@ -12,12 +12,8 @@ import pytest
 import requests
 
 from common import OperatingSystem
-from infection_monkey.exploit.agent_binary_request import (
-    AgentBinaryHTTPRequestHandler,
-    AgentBinaryRequest,
-    RequestID,
-    RequestType,
-)
+from infection_monkey.exploit.agent_binary_request import AgentBinaryRequest, RequestID, RequestType
+from infection_monkey.exploit.http_agent_binary_request_handler import AgentBinaryHTTPRequestHandler
 from infection_monkey.exploit.http_agent_binary_server import HTTPAgentBinaryServer
 from infection_monkey.network import TCPPortSelector
